fix: [ANDROAPP-6520] save complete event to Preferences from form module (#3832)
* fix: [ANDROAPP-6520] Save completed event to preferences for scheduling dialog

* fix: [ANDROAPP-6520] refactor uses and location of preference provider to repository to maintain data/ui layer separation

* fix: [ANDROAPP-6520] adapt unit tests
