+++ form/src/test/java/org/dhis2/form/ui/FormViewModelTest.kt
@@ -55,7 +55,6 @@ class FormViewModelTest {
             repository,
             dispatcher,
             geometryController,
-            preferenceProvider = preferenceProvider,
         )
     }
